pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["matthew <matthew@cupist.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
flasgger = "^0.9.4"
pytest = "^5.4.1"
alembic = "^1.4.2"
flask_sqlalchemy = "^2.4.1"
mysql-connector-python = "^8.0.19"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
